1. Updated high temperature UI Print 2. Fixed gesture questions not getting updated when changed in portal for existing workflow 3. Updated check for the institution id

+++ app/src/main/java/com/certify/snap/controller/DatabaseController.java
@@ -437,10 +437,12 @@ public class DatabaseController {
     private void deleteMemberData() {
         String path = Environment.getExternalStorageDirectory() + "/pic";
         File file = new File(path);
-        String[] children = file.list();
-        for (String child : children) {
-            new File(file, child).delete();
+        if (file.isDirectory() && file.exists()) {
+            String[] children = file.list();
+            for (String child : children) {
+                new File(file, child).delete();
+            }
+            file.delete();
         }
-        file.delete();
     }
 }
